WIP (refactor): resolve package names collision
