+++ main.py
@@ -3,7 +3,7 @@ from dotenv import load_dotenv
 from telegram.ext import ApplicationBuilder
 from state.bot_state import bot_state
 from services.watchdog import watchdog_state
-from telegram.handlers import register_handlers
+from handlers.handlers import register_handlers
 
 # Enable logging
